+++ src/app/_layout.tsx
@@ -305,7 +305,7 @@ export default function RootLayout() {
                 <Stack.Screen name="settings/language" />
                 <Stack.Screen name="settings/font" />
                 <Stack.Screen name="settings/personalization" />
-                <Stack.Screen name="settings/explore-chips" />
+                <Stack.Screen name="settings/home-chips" />
                 <Stack.Screen name="settings/song-menu" />
                 <Stack.Screen name="settings/home-sections" />
                 <Stack.Screen name="settings/equalizer" />